requirements.txt
google-generativeai>=0.7.0
langgraph>=0.1.0
manim>=0.18.0
gTTS>=2.5.0
ffmpeg-python>=0.2.0
